Refactored AlphaKeyboard to own file, added long press gesture recognizer to DualKeyboard
